LPS-196687 Rename collection-configuration to collection_configuration
